requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
numpy==2.1.3
faiss-cpu==1.9.0.post1
sentence-transformers==3.3.1
pypdf==5.1.0
openai==1.57.0
python-dotenv==1.0.1
pytest==8.3.4
httpx==0.28.1
